+++ src/types.ts
@@ -4,6 +4,7 @@ import { DataQuery } from '@grafana/schema';
 export interface RabbitMQQuery extends DataQuery { }
 
 export interface StreamOptions {
+  shouldDisposeStream: boolean;
   streamName: string;
   consumerName: string;
   offsetFromStart: boolean;
@@ -35,6 +36,8 @@ export interface RabbitMQDataSourceOptions extends DataSourceJsonData {
 }
 
 export interface ExchangeOptions {
+  shouldDisposeExchange: boolean;
+  disposeIfUnused: boolean;
   name: string;
   type: string;
   durable: boolean;
@@ -46,6 +49,7 @@ export interface ExchangeOptions {
 export interface ExchangesOptions extends Array<ExchangeOptions> {}
 
 export interface BindingOptions {
+  shouldDisposeBinding: boolean;
   isQueueBinding: boolean;
   senderName: string;
   routingKey: string;